Sanitize ELF Sections API
ELF Sections are kinda 1-indexed in that while there is a Section 0,
it is always of type SHT_NULL and is otherwise invalid to reference.
Further, later ELF extensions have re-used these unused bits for
extensions, notably for ELF files with more than 0xff00 sections.
This patch does the following:

- Prevents section 0 from being returned by "Sections", since it isn't
  actually a real section and treating it as such would be incorrect
  if the section header stores extension data (before it just doesn't
  make a difference)
- As a result indices into `Sections` now actually match the ELF section
  number rather than being off by one.
- For good measure, implement the ELF large sections extension and add
  a test to make sure we can handle a file with as many sections.